refactor: reorganise action buttons, fixed labels for translations
