+++ src/components/layout-panel/bottom-bar/action-buttons/styles/action-buttons.module.css
@@ -1,0 +1,52 @@
+.button {
+    display: flex;
+    align-items: center;
+    gap: 4px;
+    min-block-size: 28px;
+    border: 1px solid var(--colors-grey300);
+    border-radius: 3px;
+    font-size: 13px;
+    font-weight: 400;
+    line-height: 1;
+    cursor: pointer;
+    padding: 0 var(--spacers-dp8);
+    color: var(--colors-grey700);
+    background-color: var(--colors-grey-050);
+    box-shadow: none;
+}
+
+.button:disabled,
+.button.update:disabled {
+    background-color: transparent;
+    color: var(--colors-grey500);
+    cursor: not-allowed;
+    font-weight: 400;
+    border-color: var(--colors-grey300);
+}
+
+.button:hover:not(:disabled) {
+    background-color: var(--colors-grey100);
+    border-color: var(--colors-grey400);
+    color: var(--colors-grey900);
+}
+
+.button:active:not(:disabled) {
+    background-color: #d1e8fa;
+    border-color: #c8e4f9;
+    transform: translateY(1px);
+}
+
+.button.update {
+    border: 1px solid #c8e4f9;
+    padding: 0 8px 0 6px;
+    color: var(--colors-blue800);
+    background-color: var(--colors-blue100);
+    box-shadow: 0 1px 2px #2129340a;
+    font-weight: 500;
+}
+
+.button.update:hover:not(:disabled) {
+    background-color: #daedfb;
+    border-color: #c8e4f9;
+    color: var(--colors-blue900);
+}
