+++ src/components/layout-panel/bottom-bar/action-buttons/use-action-button.ts
@@ -1,236 +1,21 @@
 import i18n from '@dhis2/d2-i18n'
-import { IconSync16, Tooltip } from '@dhis2/ui'
-import cx from 'classnames'
-import { useMemo, type FC } from 'react'
-import classes from './styles/bottom-bar.module.css'
-import {
-    useAppSelector,
-    useAppDispatch,
-    useMetadataItem,
-    useMetadataStore,
-} from '@hooks'
+import { useMemo } from 'react'
+import type { ButtonAction } from './base-button'
+import { useAppSelector, useMetadataItem, useMetadataStore } from '@hooks'
 import {
     isDataSourceProgramWithoutRegistration,
     isDataSourceProgramWithRegistration,
-    isDataSourceTrackedEntityType,
 } from '@modules/data-source'
 import { isDimensionInLayout } from '@modules/layout'
 import { isVisualizationEmpty } from '@modules/visualization'
 import { getCurrentVis } from '@store/current-vis-slice'
 import { getDataSourceId } from '@store/dimensions-selection-slice'
-import { tUpdateCurrentVisFromVisUiConfig } from '@store/thunks'
 import {
     getVisUiConfigLayout,
     getVisUiConfigOutputType,
     getVisUiConfigVisualizationType,
-    setVisUiConfigOutputType,
 } from '@store/vis-ui-config-slice'
 import type { OutputType } from '@types'
-
-type ButtonAction = 'create' | 'switch' | 'update'
-
-type BaseButtonProps = {
-    action: ButtonAction
-    disabled?: boolean
-    label: string
-    tooltipProps?: object
-    type: OutputType
-}
-
-const BaseButton: FC<BaseButtonProps> = ({
-    action,
-    disabled = false,
-    label,
-    tooltipProps,
-    type,
-}) => {
-    const dispatch = useAppDispatch()
-
-    const onClick = () => {
-        dispatch(setVisUiConfigOutputType(type))
-
-        dispatch(tUpdateCurrentVisFromVisUiConfig())
-    }
-
-    return (
-        <button
-            type="button"
-            onClick={onClick}
-            disabled={disabled}
-            className={cx(classes.button, {
-                [classes.disabled]: disabled,
-                [classes.update]: action === 'update',
-            })}
-            {...tooltipProps}
-        >
-            {action === 'update' && <IconSync16 />}
-            {label}
-        </button>
-    )
-}
-
-const ActionButtonWithConditionalTooltip: FC<
-    BaseButtonProps & {
-        tooltipConfig?: { content: string; openDelay?: number }
-    }
-> = ({ tooltipConfig, ...props }) => {
-    if (tooltipConfig) {
-        const { content, openDelay = 500 } = tooltipConfig
-
-        return (
-            <Tooltip content={content} openDelay={openDelay}>
-                {(tooltipProps) => (
-                    <BaseButton {...props} tooltipProps={tooltipProps} />
-                )}
-            </Tooltip>
-        )
-    }
-
-    return <BaseButton {...props} />
-}
-
-export const EventButton: FC = () => {
-    const visualizationType = useAppSelector(getVisUiConfigVisualizationType)
-
-    const { action, dataSourceMetadata, tooltipConfig } =
-        useActionButton('EVENT')
-
-    const eventLabel = useMemo(() => {
-        if (
-            isDataSourceProgramWithRegistration(dataSourceMetadata) &&
-            dataSourceMetadata.displayEventLabel
-        ) {
-            return dataSourceMetadata.displayEventLabel
-        }
-
-        return 'Event' // XXX: translations ?!
-    }, [dataSourceMetadata])
-
-    const label = useMemo(() => {
-        const targetVisualization =
-            visualizationType === 'LINE_LIST'
-                ? i18n.t(`{{ eventLabel }} list`, { eventLabel })
-                : i18n.t(`{{ eventLabel }} table`, { eventLabel })
-
-        switch (action) {
-            case 'create':
-                return i18n.t(`Create {{targetVisualization}}`, {
-                    targetVisualization,
-                })
-            case 'switch':
-                return i18n.t(`Switch to {{targetVisualization}}`, {
-                    targetVisualization,
-                })
-            case 'update':
-                return i18n.t(`Update {{targetVisualization}}`, {
-                    targetVisualization,
-                })
-        }
-    }, [action, eventLabel, visualizationType])
-
-    return (
-        <ActionButtonWithConditionalTooltip
-            action={action}
-            disabled={Boolean(tooltipConfig)}
-            label={label}
-            tooltipConfig={tooltipConfig}
-            type="EVENT"
-        />
-    )
-}
-
-export const EnrollmentButton: FC = () => {
-    const visualizationType = useAppSelector(getVisUiConfigVisualizationType)
-
-    const { action, dataSourceMetadata, tooltipConfig } =
-        useActionButton('ENROLLMENT')
-
-    const enrollmentLabel = useMemo(() => {
-        if (
-            isDataSourceProgramWithRegistration(dataSourceMetadata) &&
-            dataSourceMetadata.displayEnrollmentLabel
-        ) {
-            return dataSourceMetadata.displayEnrollmentLabel
-        }
-
-        return 'Enrollment' // XXX: translations ?!
-    }, [dataSourceMetadata])
-
-    const label = useMemo(() => {
-        const targetVisualization =
-            visualizationType === 'LINE_LIST'
-                ? i18n.t(`{{ enrollmentLabel }} list`, { enrollmentLabel })
-                : i18n.t(`{{ enrollmentLabel }} table`, { enrollmentLabel })
-
-        switch (action) {
-            case 'create':
-                return i18n.t(`Create {{targetVisualization}}`, {
-                    targetVisualization,
-                })
-            case 'switch':
-                return i18n.t(`Switch to {{targetVisualization}}`, {
-                    targetVisualization,
-                })
-            case 'update':
-                return i18n.t(`Update {{targetVisualization}}`, {
-                    targetVisualization,
-                })
-        }
-    }, [action, enrollmentLabel, visualizationType])
-
-    return (
-        <ActionButtonWithConditionalTooltip
-            action={action}
-            disabled={Boolean(tooltipConfig)}
-            label={label}
-            tooltipConfig={tooltipConfig}
-            type="ENROLLMENT"
-        />
-    )
-}
-
-export const TrackedEntityInstanceButton: FC = () => {
-    const { action, dataSourceMetadata, tooltipConfig } = useActionButton(
-        'TRACKED_ENTITY_INSTANCE'
-    )
-
-    const trackedEntityTypeName = useMemo(() => {
-        if (isDataSourceProgramWithRegistration(dataSourceMetadata)) {
-            return dataSourceMetadata.trackedEntityType.name
-        } else if (isDataSourceTrackedEntityType(dataSourceMetadata)) {
-            return dataSourceMetadata.name
-        } else {
-            return 'tracked entity' // XXX: what about translations?
-        }
-    }, [dataSourceMetadata])
-
-    const label = useMemo(() => {
-        switch (action) {
-            case 'create':
-                return i18n.t(`Create {{trackedEntityTypeName}} list`, {
-                    trackedEntityTypeName,
-                })
-            case 'switch':
-                return i18n.t(`Switch to {{trackedEntityTypeName}} list`, {
-                    trackedEntityTypeName,
-                })
-            case 'update':
-                return i18n.t(`Update {{trackedEntityTypeName}} list`, {
-                    trackedEntityTypeName,
-                })
-        }
-    }, [action, trackedEntityTypeName])
-
-    return (
-        <ActionButtonWithConditionalTooltip
-            action={action}
-            disabled={Boolean(tooltipConfig)}
-            label={label}
-            tooltipConfig={tooltipConfig}
-            type="TRACKED_ENTITY_INSTANCE"
-        />
-    )
-}
 
 type TooltipContent = { content: string; openDelay?: number } | undefined
